"Initial"
Forgot to repo this for a while. Currently functional and quite well
tested.

Still lots to do, considering an extensive re-write using a more
function-oriented style (avoiding the statefulness associated with
hoisting variables from apply-functions. Not necessarily so bad, but i
don't think it makes for pretty code!)

This library is being used in production @ PRIO (2019)